@@ -1,53 +1,41 @@
 {
   "name": "ostips",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod --build-optimizer --base-href /ostips/",
-    "after-deploy": "sh scripts/afterDeploy.sh",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.0.0-next.10",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "~9.0.0-next.10",
-    "@angular/compiler": "~9.0.0-next.10",
-    "@angular/core": "~9.0.0-next.10",
-    "@angular/forms": "~9.0.0-next.10",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "~9.0.0-next.10",
-    "@angular/platform-browser-dynamic": "~9.0.0-next.10",
-    "@angular/pwa": "^0.803.9",
-    "@angular/router": "~9.0.0-next.10",
-    "@angular/service-worker": "~9.0.0-next.10",
-    "hammerjs": "^2.0.8",
-    "rxjs": "~6.5.3",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "@angular/animations": "^18.1.0",
+    "@angular/cdk": "18.2.10",
+    "@angular/common": "^18.1.0",
+    "@angular/compiler": "^18.1.0",
+    "@angular/core": "^18.1.0",
+    "@angular/forms": "^18.1.0",
+    "@angular/material": "18.2.10",
+    "@angular/platform-browser": "^18.1.0",
+    "@angular/platform-browser-dynamic": "^18.1.0",
+    "@angular/router": "^18.1.0",
+    "@angular/service-worker": "^18.1.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.802.1",
-    "@angular/cli": "~8.2.1",
-    "@angular/compiler-cli": "~9.0.0-next.10",
-    "@angular/language-service": "~9.0.0-next.10",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "@angular-devkit/build-angular": "^18.1.2",
+    "@angular/cli": "^18.1.2",
+    "@angular/compiler-cli": "^18.1.0",
+    "@types/jasmine": "~5.1.0",
+    "jasmine-core": "~5.1.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.5.2"
   }
 }
